# apps/backend/tests/openAPI/ragie.json
{
    "openapi": "3.1.0",
    "info": { "title": "Ragie API", "version": "1.0.0" },
    "servers": [{ "url": "https://api.ragie.ai" }],
    "paths": {
...
        "/documents/{document_id}": {
            "get": {
                "tags": ["documents"],
                "summary": "Get Document",
                "operationId": "GetDocument",
                "security": [{ "auth": [] }],
                "parameters": [
                    {
                        "name": "document_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "description": "The id of the document.",
                            "examples": ["00000000-0000-0000-0000-000000000000"],
                            "title": "Document Id"
                        },
                        "description": "The id of the document."
                    },
                    {
                        "name": "partition",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [{ "type": "string" }, { "type": "null" }],
                            "description": "An optional partition to scope the request to. If omitted, accounts created after 1/9/2025 will have the request scoped to the default partition, while older accounts will have the request scoped to all partitions. Older accounts may opt in to strict partition scoping by contacting support@ragie.ai. Older accounts using the partitions feature are strongly recommended to scope the request to a partition.",
                            "examples": ["acme_customer_id"],
                            "title": "Partition"
                        },
                        "description": "An optional partition to scope the request to. If omitted, accounts created after 1/9/2025 will have the request scoped to the default partition, while older accounts will have the request scoped to all partitions. Older accounts may opt in to strict partition scoping by contacting support@ragie.ai. Older accounts using the partitions feature are strongly recommended to scope the request to a partition."
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DocumentGet" } } }
                    },
                    "401": {
                        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorMessage" } } },
                        "description": "Unauthorized"
                    },
                    "402": {
                        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorMessage" } } },
                        "description": "Payment Required"
                    },
                    "404": {
                        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorMessage" } } },
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
                    },
                    "429": {
                        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorMessage" } } },
                        "description": "Too Many Requests"
                    }
                },
                "x-speakeasy-name-override": "get"
            },
            "delete": {
                "tags": ["documents"],
                "summary": "Delete Document",
                "operationId": "DeleteDocument",
                "security": [{ "auth": [] }],
                "parameters": [
                    {
                        "name": "document_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "description": "The id of the document.",
                            "examples": ["00000000-0000-0000-0000-000000000000"],
                            "title": "Document Id"
                        },
                        "description": "The id of the document."
                    },
                    {
                        "name": "async",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [{ "type": "boolean" }, { "type": "null" }],
                            "description": "If true, performs document deletion asynchronously",
                            "examples": [false],
                            "default": false,
                            "title": "Async"
                        },
                        "description": "If true, performs document deletion asynchronously"
                    },
                    {
                        "name": "partition",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [{ "type": "string" }, { "type": "null" }],
                            "description": "An optional partition to scope the request to. If omitted, accounts created after 1/9/2025 will have the request scoped to the default partition, while older accounts will have the request scoped to all partitions. Older accounts may opt in to strict partition scoping by contacting support@ragie.ai. Older accounts using the partitions feature are strongly recommended to scope the request to a partition.",
                            "examples": ["acme_customer_id"],
                            "title": "Partition"
                        },
                        "description": "An optional partition to scope the request to. If omitted, accounts created after 1/9/2025 will have the request scoped to the default partition, while older accounts will have the request scoped to all partitions. Older accounts may opt in to strict partition scoping by contacting support@ragie.ai. Older accounts using the partitions feature are strongly recommended to scope the request to a partition."
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DocumentDelete" } } }
                    },
                    "202": {
                        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DocumentDelete" } } },
                        "description": "Accepted"
                    },
                    "401": {
                        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorMessage" } } },
                        "description": "Unauthorized"
                    },
                    "402": {
                        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorMessage" } } },
                        "description": "Payment Required"
                    },
                    "404": {
                        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorMessage" } } },
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
                    },
                    "429": {
                        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorMessage" } } },
                        "description": "Too Many Requests"
                    }
                },
                "x-speakeasy-name-override": "delete"
            }
        },
...
        "/retrievals": {
            "post": {
                "tags": ["retrievals"],
                "summary": "Retrieve",
                "operationId": "Retrieve",
                "requestBody": {
                    "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RetrieveParams" } } },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Retrieval" } } }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorMessage" } } }
                    },
                    "402": {
                        "description": "Payment Required",
                        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorMessage" } } }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorMessage" } } }
                    }
                },
                "security": [{ "auth": [] }],
                "x-speakeasy-name-override": "retrieve"
            }
        },
...
    },
...
    "components": {
        "schemas": {
...
            "DocumentDelete": {
                "properties": { "status": { "type": "string", "title": "Status" } },
                "type": "object",
                "required": ["status"],
                "title": "DocumentDelete"
            },
...
            "DocumentGet": {
                "properties": {
                    "status": { "type": "string", "title": "Status" },
                    "id": { "type": "string", "format": "uuid", "title": "Id" },
                    "created_at": { "type": "string", "format": "date-time", "title": "Created At" },
                    "updated_at": { "type": "string", "format": "date-time", "title": "Updated At" },
                    "name": { "type": "string", "title": "Name" },
                    "metadata": {
                        "additionalProperties": {
                            "anyOf": [
                                { "type": "string" },
                                { "type": "integer" },
                                { "type": "boolean" },
                                { "items": { "type": "string" }, "type": "array" }
                            ]
                        },
                        "type": "object",
                        "title": "Metadata"
                    },
                    "partition": { "type": "string", "title": "Partition" },
                    "chunk_count": { "anyOf": [{ "type": "integer" }, { "type": "null" }], "title": "Chunk Count" },
                    "external_id": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "External Id" },
                    "page_count": { "anyOf": [{ "type": "number" }, { "type": "null" }], "title": "Page Count" },
                    "errors": { "items": { "type": "string" }, "type": "array", "title": "Errors" }
                },
                "type": "object",
                "required": ["status", "id", "created_at", "updated_at", "name", "metadata", "partition", "errors"],
                "title": "DocumentGet"
            },
...
            "ErrorMessage": {
                "properties": { "detail": { "type": "string", "title": "Detail" } },
                "type": "object",
                "required": ["detail"],
                "title": "ErrorMessage"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": { "items": { "$ref": "#/components/schemas/ValidationError" }, "type": "array", "title": "Detail" }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "Retrieval": {
                "properties": {
                    "scored_chunks": { "items": { "$ref": "#/components/schemas/ScoredChunk" }, "type": "array", "title": "Scored Chunks" }
                },
                "type": "object",
                "required": ["scored_chunks"],
                "title": "Retrieval"
            },
...
        },
        "securitySchemes": { "auth": { "type": "http", "scheme": "bearer" } }
    },
    "security": [{ "auth": [] }],
...
}